+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Full changelog for ZF3 view helpers library.
 
+## v0.20 - Buttons - 2017-05-19
+
+* Added Bootstrap 4 button view helper.
+
+## v0.10.1 - Icons - 2017-05-16
+
+* Added missing icons to README.
+
 ## v0.10 (Initial release) - 2017-05-16
 
 * Added Bootstrap 3 Jumbotron view helper.